+++ .jshintrc
@@ -1,0 +1,4 @@
+{
+    // Bad line breaking
+    "laxbreak": true
+}
