+++ examples/example.html
@@ -30,7 +30,7 @@
     <canvas id="videoCanvas" width="320" height="180"></canvas>
 </div>
 
-<script type="text/javascript" src="./jsmpg.js"></script>
+<script type="text/javascript" src="../bin/jsmpg.js"></script>
 <script type="text/javascript">
     var canvas = document.getElementById('videoCanvas');
