+++ examples/stream-example.html
@@ -30,7 +30,7 @@
 			<a href="http://www.apple.com/safari/">Safari</a> or Internet Explorer 10
 		</p>
 	</canvas>
-	<script type="text/javascript" src="jsmpg.js"></script>
+	<script type="text/javascript" src="../bin/jsmpg.js"></script>
 	<script type="text/javascript">
 		// Setup the WebSocket connection and start the player
 		var client = new WebSocket( 'ws://example.com:8084/' );
